Created GameController class and moved all game code to it
